App split into components
