+++ frontend/app/app.module.ts
@@ -4,10 +4,18 @@ import { FormsModule }   from '@angular/forms';
 import { MomentModule }    from 'angular2-moment';
 
 import { AppComponent }  from './app.component';
+import { TimePickerComponent }  from './time-pick.component';
+import { DevListComponent }  from './dev-list.component';
+import { CompanyListComponent }  from './company-list.component';
 
 @NgModule({
   imports:      [ BrowserModule, FormsModule, MomentModule ],
-  declarations: [ AppComponent ],
+  declarations: [
+	  AppComponent,
+	  TimePickerComponent,
+	  DevListComponent,
+	  CompanyListComponent
+  ],
   bootstrap: [ AppComponent ]
 })
